+++ src/app/config/app.constants.ts
@@ -25,7 +25,7 @@ Bug Fixes:
     DROPBOX: 'kbwh5504okeu3b9',
     GOOGLE_DRIVE: '801129580490-kbquipjice5v26b5jtpmhnncs31li7pv.apps.googleusercontent.com',
   },
-  DATE_FORMAT: 'MM/DD/YYYY',
+  LONG_DATE_FORMAT: 'L',
   QUOTE_CACHE_TIMEOUT: 60 * 60 * 11, // value in seconds. do not request updated quotes for assets until timeout
   QUOTE_SERVICE_BASE_URL: 'https://roadtofire.iliviu.me:8443/',
   SUPPORTED_EXCHANGES: {