+++ src/app/core/services/rs-app-storage.service.ts
@@ -66,6 +66,9 @@ function createAppRStorageModule(serializer: StorageSerializer, remoteStorage: R
           'encryptedData': {
             'type': 'string',
           },
+          'dateAndCurrencyFormat': {
+            'type': 'string'
+          },
           'saveOnCloud': {
             'type': 'boolean'
           },