+++ src/app/config/app.constants.ts
@@ -11,15 +11,13 @@ export interface SymbolDetails {
 
 export const APP_CONSTS = {
   TITLE: 'Road To FIRE',
-  VERSION: '1.0.2',
+  VERSION: '1.1.0',
   // text to show when user first loads the current new version
-  WHATS_NEW_TITLE: `What's new in version 1.0.2`,
+  WHATS_NEW_TITLE: `What's new in version 1.1.0`,
   WHATS_NEW:
-    `UI Changes:
-* Display tooltips of nearest point when hovering history charts, not only when hovering the exact point location.
-
-Bug Fixes:
-* Correctly set currency symbol when a cash asset is selected by default in asset trade dialog`,
+    `Features:
+  * Add option to select from multiple date and currency formats.
+  * Display loan-to-value ratio in account overview tab.`,
   REPOSITORY_URL: 'https://github.com/iLiviu/road-to-fire',
   STORAGE_API_KEYS: {
     DROPBOX: 'kbwh5504okeu3b9',