+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Master
+
+### Features
+* Add option to select from multiple date and currency formats.
+
 ## 1.0.2 (2020-01-15)
 
 ### UI Changes