+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Master
+## 1.1.0 (2020-01-27)
 
 ### Features
 * Add option to select from multiple date and currency formats.